backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
flask-mail==0.9.1
python-dotenv==1.0.1
google-auth==2.23.0
psycopg2-binary>=2.9.9
gunicorn==21.2.0
